backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0

python-dotenv==1.0.0
gunicorn==22.0.0
gevent==24.2.1
requests==2.31.0

numpy==1.24.3

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.0
